requirements.txt
streamlit
python-dotenv
google-generativeai
beautifulsoup4
requests
PyMuPDF
python-docx
fpdf